backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-dotenv
requests
bytez
Pillow